Got the controls looking okay
It would be nice to position the Rotation amount relative to the slider.  An
idea would be to have the bounding box for the rotation amount be fixed and the
text always centred, because we sort of have an idea of the largest number
(360.00) that could fit in the box.